fix refetch
